Fixed bugs and added help

+++ src/commands/clear.js
@@ -3,24 +3,37 @@ const { CommandOptions } = require("../structures")
 
 class Clear extends Command {
   constructor(client) {
-		const options = new CommandOptions(`clear`, 1000, 2, 1)
-    super(client, options)
+
+		const requiredPerms = [`MANAGE_MESSAGES`]
+		const options = new CommandOptions(
+			`clear`,
+			1000,
+			1,
+			requiredPerms,
+			`بتمسح رسايل بعدد n`,
+			`val! clear 5`
+		)
+
+		super(client, options)
   }
 
-  async run(context) {
+  async _run(context) {
 		const { message, params, channel } = context
-		const count = parseInt(params[0])
+		const numbersRegex = /\d+/
 
-		if(count === 0) {
-			message.reply('لما تكتب صفر للكوماند دي هتخلي ديسكورد يمسح كل الرسايل اللي ف التشانل! جرب رقم تاني')
-			return
+		if(numbersRegex.test(params[0])){
+			const count = parseInt(params[0])
+
+			if(count === 0) {
+				message.reply('لما تكتب صفر للكوماند دي هتخلي ديسكورد يمسح كل الرسايل اللي ف التشانل! جرب رقم تاني')
+				return
+			}
+
+			await channel.bulkDelete(count)
+			message.reply(`مسحت ${count} رسايل. تحب اجيبلك كوبايتين لمون؟`)
+
 		}
-
-		await channel.bulkDelete(count)
-		// deletedMessages = deletedMessages.map(message => ({ author: { name: message.author.username, id: message.author.id }, content: message.content }))
-		// __ENV.__DATABASE_OBJECT.collection('DELETED_MESSAGES').insertMany(deletedMessages)
-
-		message.reply(`مسحت ${count} رسايل. تحب اجيبلك كوبايتين لمون؟`)
+		else message.reply(`لازم تدخل رقم كـتالت باراميتير للكوماند دي`)
 	}
 
 }
